requirements.txt
anthropic>=0.40.0
Pillow>=10.0
matplotlib>=3.8
numpy>=1.26
pandas>=2.0
tqdm>=4.66
datasets>=2.16
streamlit>=1.30
torch>=2.0
transformers>=4.36
